refactor(sprite)!: experiment with another structure
